ArrayIndexing: Split and move related tests for array_indexing
