+++ loki/transformations/array_indexing/tests/test_array_promote.py
@@ -1,0 +1,142 @@
+# (C) Copyright 2018- ECMWF.
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+import pytest
+import numpy as np
+
+from loki import Subroutine
+from loki.jit_build import jit_compile
+from loki.expression import symbols as sym
+from loki.frontend import available_frontends
+
+from loki.transformations.array_indexing.promote import promote_variables
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_transform_promote_variable_scalar(tmp_path, frontend):
+    """
+    Apply variable promotion for a single scalar variable.
+    """
+    fcode = """
+subroutine transform_promote_variable_scalar(ret)
+  implicit none
+  integer, intent(out) :: ret
+  integer :: tmp, jk
+
+  ret = 0
+  do jk=1,10
+    tmp = jk
+    ret = ret + tmp
+  end do
+end subroutine transform_promote_variable_scalar
+    """.strip()
+    routine = Subroutine.from_source(fcode, frontend=frontend)
+
+    # Test the original implementation
+    filepath = tmp_path/(f'{routine.name}_{frontend}.f90')
+    function = jit_compile(routine, filepath=filepath, objname=routine.name)
+    ret = function()
+    assert ret == 55
+
+    # Apply and test the transformation
+    assert isinstance(routine.variable_map['tmp'], sym.Scalar)
+    promote_variables(routine, ['TMP'], pos=0, index=routine.variable_map['JK'], size=sym.Literal(10))
+    assert isinstance(routine.variable_map['tmp'], sym.Array)
+    assert routine.variable_map['tmp'].shape == (sym.Literal(10),)
+
+    promoted_filepath = tmp_path/(f'{routine.name}_promoted_{frontend}.f90')
+    promoted_function = jit_compile(routine, filepath=promoted_filepath, objname=routine.name)
+    ret = promoted_function()
+    assert ret == 55
+
+
+@pytest.mark.parametrize('frontend', available_frontends())
+def test_transform_promote_variables(tmp_path, frontend):
+    """
+    Apply variable promotion for scalar and array variables.
+    """
+    fcode = """
+subroutine transform_promote_variables(scalar, vector, n)
+  implicit none
+  integer, intent(in) :: n
+  integer, intent(inout) :: scalar, vector(n)
+  integer :: tmp_scalar, tmp_vector(n), tmp_matrix(n,n)
+  integer :: jl, jk
+
+  do jl=1,n
+    ! a bit of a hack to create initialized meaningful output
+    tmp_vector(:) = 0
+  end do
+
+  do jl=1,n
+    tmp_scalar = jl
+    tmp_vector(jl) = jl
+
+    do jk=1,n
+      tmp_matrix(jk, jl) = jl + jk
+    end do
+  end do
+
+  scalar = 0
+  do jl=1,n
+    scalar = scalar + tmp_scalar
+    vector = tmp_matrix(:,jl) + tmp_vector(:)
+  end do
+end subroutine transform_promote_variables
+    """.strip()
+    routine = Subroutine.from_source(fcode, frontend=frontend)
+
+    # Test the original implementation
+    filepath = tmp_path/(f'{routine.name}_{frontend}.f90')
+    function = jit_compile(routine, filepath=filepath, objname=routine.name)
+
+    n = 10
+    scalar = np.array(0)
+    vector = np.zeros(shape=(n,), order='F', dtype=np.int32)
+    function(scalar, vector, n)
+    assert scalar == n*n
+    assert np.all(vector == np.array(list(range(1, 2*n+1, 2)), order='F', dtype=np.int32) + n + 1)
+
+    # Verify dimensions before promotion
+    assert isinstance(routine.variable_map['tmp_scalar'], sym.Scalar)
+    assert isinstance(routine.variable_map['tmp_vector'], sym.Array)
+    assert routine.variable_map['tmp_vector'].shape == (routine.variable_map['n'],)
+    assert isinstance(routine.variable_map['tmp_matrix'], sym.Array)
+    assert routine.variable_map['tmp_matrix'].shape == (routine.variable_map['n'], routine.variable_map['n'])
+
+    # Promote scalar and vector and verify dimensions
+    promote_variables(routine, ['tmp_scalar', 'tmp_vector'], pos=-1, index=routine.variable_map['JL'],
+                      size=routine.variable_map['n'])
+
+    assert isinstance(routine.variable_map['tmp_scalar'], sym.Array)
+    assert routine.variable_map['tmp_scalar'].shape == (routine.variable_map['n'],)
+    assert isinstance(routine.variable_map['tmp_vector'], sym.Array)
+    assert routine.variable_map['tmp_vector'].shape == (routine.variable_map['n'], routine.variable_map['n'])
+    assert isinstance(routine.variable_map['tmp_matrix'], sym.Array)
+    assert routine.variable_map['tmp_matrix'].shape == (routine.variable_map['n'], routine.variable_map['n'])
+
+    # Promote matrix and verify dimensions
+    promote_variables(routine, ['tmp_matrix'], pos=1, index=routine.variable_map['JL'],
+                      size=routine.variable_map['n'])
+
+    assert isinstance(routine.variable_map['tmp_scalar'], sym.Array)
+    assert routine.variable_map['tmp_scalar'].shape == (routine.variable_map['n'],)
+    assert isinstance(routine.variable_map['tmp_vector'], sym.Array)
+    assert routine.variable_map['tmp_vector'].shape == (routine.variable_map['n'], routine.variable_map['n'])
+    assert isinstance(routine.variable_map['tmp_matrix'], sym.Array)
+    assert routine.variable_map['tmp_matrix'].shape == (routine.variable_map['n'], ) * 3
+
+    # Test promoted routine
+    promoted_filepath = tmp_path/(f'{routine.name}_promoted_{frontend}.f90')
+    promoted_function = jit_compile(routine, filepath=promoted_filepath, objname=routine.name)
+
+    scalar = np.array(0)
+    vector = np.zeros(shape=(n,), order='F', dtype=np.int32)
+    promoted_function(scalar, vector, n)
+    assert scalar == n*(n+1)//2
+    assert np.all(vector[:-1] == np.array(list(range(n + 1, 2*n)), order='F', dtype=np.int32))
+    assert vector[-1] == 3*n
